+++ src/lib.rs
@@ -734,6 +734,38 @@ pub fn rewrite_url<'a>(config: &Config, url: &'a Url) -> Result<Cow<'a, Url>> {
     }
 }
 
+/// Walks a given storage URL as if it were a directory.
+///
+/// Returns a list of relative paths from the given URL.
+///
+/// If the given storage URL is not a directory, an empty list is returned.
+pub async fn walk(config: Config, client: HttpClient, mut url: Url) -> Result<Vec<String>> {
+    if let Ok(mut segments) = url.path_segments_mut() {
+        // Push an empty segment to treat the URL as a directory
+        // This ensures there is no leading slash in the returned relative paths.
+        segments.pop_if_empty().push("");
+    }
+
+    if AzureBlobStorageBackend::is_supported_url(&config, &url) {
+        let url = AzureBlobStorageBackend::rewrite_url(&config, &url)?;
+        AzureBlobStorageBackend::new(config, client, None)
+            .walk(url.into_owned())
+            .await
+    } else if S3StorageBackend::is_supported_url(&config, &url) {
+        let url = S3StorageBackend::rewrite_url(&config, &url)?;
+        S3StorageBackend::new(config, client, None)
+            .walk(url.into_owned())
+            .await
+    } else if GoogleStorageBackend::is_supported_url(&config, &url) {
+        let url = GoogleStorageBackend::rewrite_url(&config, &url)?;
+        GoogleStorageBackend::new(config, client, None)
+            .walk(url.into_owned())
+            .await
+    } else {
+        Err(Error::UnsupportedUrl(url))
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;